Basic immune factors and HLA factors working
Implemented HLA A, B and C with recognized supertypes (note: supertypes
containing important subtypes are set to approximate those subtypes,
supertypes without an important subtype do not have an accurate rate.
Supertypes with an important subtype are set to approximately that
subtype's rate  in a Caucasian population.) Gene class contains link
references and values for the particular gene effect. This means that
there's only one location to change for such effects and simplifies
program enhancements.