Added Edge Tracking completed upgrades
Moved most functions that had been moved into the "male" "female"
classes back into agent since the main reason why they were in there was
because of the "get male" "get female" options in relationship.
Finished moving creation of and processing of networks into other
steppables. Added checks for homosexual coital infection to males, all
males now defaulting to circumcised. Added some check methods to disease
matrix and Agent to make things go smoother and remove a lot of
redundant checks.
Streamlined remove edge.
Handled the "edge mapping" with the RelationshipNetwork class. Note that
this class may become "not useful" and replaced with an arraylist for
nodes in the future. It's mains reason for being is that the Network
class produces the visualizations for the network which helps with
debugging, when the sim is ready to be run without visualizations we can
change it.
Might have missed something, but you can check the diff if you're
interested.