Homosexual transmission set up, rebalanced network creation, debugged
Fixed up relationship creation- Relationships require both parties to
agree to the relationship. Marriages only occur after relationships, not
right off the bat.
Homosexual transmission is setup, however, since I'm not sure how to
assign or handle varying degrees of sexuality I am leaving the
implementation out until I can discuss it with a specialist.
Handled the "custom network" issue with a work around while we still
need the visual network- there is a second set of edges currently being
stored (addresses only) in a linked list that is "wrapped" around the
relationship network.
Note that this may be changed in the next step which is to create
neighborhoods.