Added a second network for MSM, upgraded logger
There was an issue with msm not being able to find other msm partners,
so I added an additional network for msm. This network can be
deactivated to require msm to continue sampling from the list of all men
and women. MSMW will randomly select the from one network or the other
each tick.