Added Testing, upgraded analysis
Added testing for our agents. Currently the only change is +25% chance
of using a condom, however it might be a basis for deterministic usage
of PrEP.
All of the defaults are set to about midline, but the best transmission
is when monogamy and commitment are <5.
Expanded the Simulation Analysis with sqldf. I'm sure there is a very R
way to do it.... I just don't know that very R way, so yay sql! I also
expanded the documentation a bit now that the changes should be
gradually becoming more minor...