Death by AIDS correction, libido network level cast to int, updated Analysis
-Death by AIDS created eternally counted agents for population growth
which would cause an unrealistic increase in new agents after enough
deaths occurred. Now they last only as long as they normally would
"live".
- Libido vs network value calculations had been calculated as double to
double, casting to int. A size difference of less than 1 does not cause
an agent to go looking or accept new connections
- There was an error running the analysis when there were no AIDS deaths
or positive tests. Added if then else statements to the prints that are
sensitive to nulls.