Updated Index to MASON std, Moved adjustlack to agent, added remove the dead option
The index file hadn't been in the right format for MASON... changing
this TO the right format doesn't seem to have helped any with the About
tab not being filled in.
Adjust Lack was being handled both in the network processing and in the
agent, moved that to just in the agent to avoid issues such as 6+3-4 = 3
(lack caps at 7) The result should be 5. This makes a big difference
with the movement of lack and was a bug.
The model can now choose if it wants to keep those who died of HIV on
the counter for population growth.